file_sharing_app/backend/requirement.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-Mail
load_dotenv